Cleanup legacy scripts
