+++ package.json
@@ -4,18 +4,12 @@
   "description": "[Check out our Contribution File](https://github.com/axa-ch/patterns-library/blob/develop/CONTRIBUTION.md#rules-and-lintings)",
   "module": "lib",
   "scripts": {
-    "postinstall": "npx lerna bootstrap --hoist && npm run build-materials",
+    "build": "NODE_ENV=production npm run build-components",
+    "watch": "npm run watch-components",
     "build-components": "npx lerna run build --parallel",
     "watch-components": "npx lerna run watch --parallel",
     "setup-npm-links": "node scripts/setup-npm-link.js",
-    "build": "npm run clean-build && npm run build-materials && cross-env NODE_ENV=production rollup --config ./rollup.config.js && node scripts/build.finalise.js",
-    "build-materials": "npm run build-icons && node scripts/build.materials.js",
-    "build-icons": "npm run clean-materials && svgo --disable=removeViewBox -f src/components/00-materials/icons -o src/components/00-materials/icons/.tmp && node scripts/build.icon-set.js && npm run clean-tmp-icons",
-    "clean-build": "npm run clean-build-tmp && rimraf src/components/*/*/lib && rimraf src/components/*/*/dist",
-    "clean-build-tmp": "rimraf src/components/*/*/.tmp-lib && rimraf src/components/*/*/.tmp-dist",
-    "clean-materials": "npm run clean-build && npm run clean-tmp-icons",
-    "clean-lib-materials": "rimraf src/components/00-materials/lib/**",
-    "clean-tmp-icons": "rimraf src/components/00-materials/icons/.tmp",
+    "postinstall": "npx lerna bootstrap --hoist && npm run build-components",
     "release": "npm whoami && npx lerna bootstrap --hoist && npm run build && npm run test && lerna publish",
     "start": "npm run storybook",
     "storybook": "start-storybook -p 6006 -c .storybook -s src/static",
